Deployment PR - 45

+++ services/apps/portal/src/application/application-owner/application-owner.controller.ts
@@ -20,9 +20,9 @@ import { DocumentService } from '../../document/document.service';
 import { AttachExternalDocumentDto } from '../application-document/application-document.dto';
 import { ApplicationService } from '../application.service';
 import {
+  APPLICATION_OWNER,
   ApplicationOwnerCreateDto,
   ApplicationOwnerDto,
-  APPLICATION_OWNER,
   ApplicationOwnerUpdateDto,
   SetPrimaryContactDto,
 } from './application-owner.dto';
@@ -76,8 +76,8 @@ export class ApplicationOwnerController {
     @Body() updateDto: ApplicationOwnerUpdateDto,
     @Req() req,
   ) {
+    await this.verifyAccessAndGetOwner(req, uuid);
     this.verifyDto(updateDto);
-    await this.ownerService.getByOwner(req.user.entity, uuid);
 
     const newParcel = await this.ownerService.update(uuid, updateDto);
 
@@ -90,7 +90,7 @@ export class ApplicationOwnerController {
 
   @Delete('/:uuid')
   async delete(@Param('uuid') uuid: string, @Req() req) {
-    const owner = await this.ownerService.getByOwner(req.user.entity, uuid);
+    const owner = await this.verifyAccessAndGetOwner(req, uuid);
     return { uuid: await this.ownerService.delete(owner) };
   }
 
@@ -100,7 +100,8 @@ export class ApplicationOwnerController {
     @Param('parcelUuid') parcelUuid: string,
     @Req() req,
   ) {
-    await this.ownerService.getByOwner(req.user.entity, uuid);
+    await this.verifyAccessAndGetOwner(req, uuid);
+
     return { uuid: await this.ownerService.attachToParcel(uuid, parcelUuid) };
   }
 
@@ -110,7 +111,8 @@ export class ApplicationOwnerController {
     @Param('parcelUuid') parcelUuid: string,
     @Req() req,
   ) {
-    await this.ownerService.getByOwner(req.user.entity, uuid);
+    await this.verifyAccessAndGetOwner(req, uuid);
+
     return { uuid: await this.ownerService.removeFromParcel(uuid, parcelUuid) };
   }
 
@@ -187,8 +189,7 @@ export class ApplicationOwnerController {
         agentOwner,
       );
     } else if (data.ownerUuid) {
-      const primaryContactOwner = await this.ownerService.getByOwner(
-        req.user.entity,
+      const primaryContactOwner = await this.ownerService.getOwner(
         data.ownerUuid,
       );
 
@@ -216,7 +217,7 @@ export class ApplicationOwnerController {
 
   @Get(':uuid/corporateSummary')
   async openCorporateSummary(@Param('uuid') uuid: string, @Req() req) {
-    const owner = await this.ownerService.getByOwner(req.user.entity, uuid);
+    const owner = await this.verifyAccessAndGetOwner(req, uuid);
 
     if (!owner.corporateSummary) {
       throw new BadRequestException('Owner has no corporate summary');
@@ -227,5 +228,15 @@ export class ApplicationOwnerController {
         owner.corporateSummary.alcsDocumentUuid,
       ),
     );
+  }
+
+  private async verifyAccessAndGetOwner(@Req() req, ownerUuid: string) {
+    const owner = await this.ownerService.getOwner(ownerUuid);
+    await this.applicationService.verifyAccess(
+      owner.applicationFileNumber,
+      req.user.entity,
+    );
+
+    return owner;
   }
 }
